refactor: move TIntMaskVerifier ops to separate class
